+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
 
 
+## 3.2.2 - 2023-02-25
+
+### Fixed
+- Fix issue with wrong tyre wear in outlaps 
+
+### Added
+- Collect ERS status for each sector
+- Collect fuel remaining for each sector
+
+
 ## 3.2.1 - 2023-02-21
 
 ### Fixed